fix: search and select asset
