+++ src/app/store/assets/asset-search.ts
@@ -1,5 +1,3 @@
 import { atom } from 'jotai';
 
 export const selectedAssetIdState = atom<string | undefined>(undefined);
-
-export const searchInputStore = atom<string>('');
